third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-all-expected.txt

Test calling getPropertyValue on computed styles for all shorthand property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("initial")).all is ""
PASS getComputedStyle(document.getElementById("inherit")).all is ""
PASS getComputedStyle(document.getElementById("invalid")).all is ""
PASS getComputedStyle(document.getElementById("notOnlyInitial")).all is ""
PASS getComputedStyle(document.getElementById("notOnlyInherit")).all is ""
PASS successfullyParsed is true

TEST COMPLETE
